[AIP-21] Move contrib.operators.s3_list_operator airflow.contrib.operators.s3_list_operator
