+++ airflow/contrib/operators/s3_list_operator.py
@@ -16,84 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""This module is deprecated. Please use `airflow.providers.amazon.aws.operators.s3_list`."""
 
-from typing import Iterable
+import warnings
 
-from airflow.models import BaseOperator
-from airflow.providers.amazon.aws.hooks.s3 import S3Hook
-from airflow.utils.decorators import apply_defaults
+# pylint: disable=unused-import
+from airflow.providers.amazon.aws.operators.s3_list import S3ListOperator  # noqa
 
-
-class S3ListOperator(BaseOperator):
-    """
-    List all objects from the bucket with the given string prefix in name.
-
-    This operator returns a python list with the name of objects which can be
-    used by `xcom` in the downstream task.
-
-    :param bucket: The S3 bucket where to find the objects. (templated)
-    :type bucket: str
-    :param prefix: Prefix string to filters the objects whose name begin with
-        such prefix. (templated)
-    :type prefix: str
-    :param delimiter: the delimiter marks key hierarchy. (templated)
-    :type delimiter: str
-    :param aws_conn_id: The connection ID to use when connecting to S3 storage.
-    :type aws_conn_id: str
-    :param verify: Whether or not to verify SSL certificates for S3 connection.
-        By default SSL certificates are verified.
-        You can provide the following values:
-
-        - ``False``: do not validate SSL certificates. SSL will still be used
-                 (unless use_ssl is False), but SSL certificates will not be
-                 verified.
-        - ``path/to/cert/bundle.pem``: A filename of the CA cert bundle to uses.
-                 You can specify this argument if you want to use a different
-                 CA cert bundle than the one used by botocore.
-    :type verify: bool or str
-
-
-    **Example**:
-        The following operator would list all the files
-        (excluding subfolders) from the S3
-        ``customers/2018/04/`` key in the ``data`` bucket. ::
-
-            s3_file = S3ListOperator(
-                task_id='list_3s_files',
-                bucket='data',
-                prefix='customers/2018/04/',
-                delimiter='/',
-                aws_conn_id='aws_customers_conn'
-            )
-    """
-    template_fields = ('bucket', 'prefix', 'delimiter')  # type: Iterable[str]
-    ui_color = '#ffd700'
-
-    @apply_defaults
-    def __init__(self,
-                 bucket,
-                 prefix='',
-                 delimiter='',
-                 aws_conn_id='aws_default',
-                 verify=None,
-                 *args,
-                 **kwargs):
-        super().__init__(*args, **kwargs)
-        self.bucket = bucket
-        self.prefix = prefix
-        self.delimiter = delimiter
-        self.aws_conn_id = aws_conn_id
-        self.verify = verify
-
-    def execute(self, context):
-        hook = S3Hook(aws_conn_id=self.aws_conn_id, verify=self.verify)
-
-        self.log.info(
-            'Getting the list of files from bucket: %s in prefix: %s (Delimiter {%s)',
-            self.bucket, self.prefix, self.delimiter
-        )
-
-        return hook.list_keys(
-            bucket_name=self.bucket,
-            prefix=self.prefix,
-            delimiter=self.delimiter)
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.operators.s3_list`.",
+    DeprecationWarning, stacklevel=2
+)
