+++ airflow/contrib/operators/s3_to_gcs_operator.py
@@ -19,10 +19,10 @@
 import warnings
 from tempfile import NamedTemporaryFile
 
-from airflow.contrib.operators.s3_list_operator import S3ListOperator
 from airflow.exceptions import AirflowException
 from airflow.gcp.hooks.gcs import GCSHook, _parse_gcs_url
 from airflow.providers.amazon.aws.hooks.s3 import S3Hook
+from airflow.providers.amazon.aws.operators.s3_list import S3ListOperator
 from airflow.utils.decorators import apply_defaults
