+++ scripts/ci/pylint_todo.txt
@@ -52,7 +52,7 @@
 ./airflow/contrib/operators/redis_publish_operator.py
 ./airflow/providers/amazon/aws/operators/s3_copy_object.py
 ./airflow/providers/amazon/aws/operators/s3_delete_objects.py
-./airflow/contrib/operators/s3_list_operator.py
+./airflow/providers/amazon/aws/operators/s3_list.py
 ./airflow/contrib/operators/s3_to_gcs_operator.py
 ./airflow/contrib/operators/s3_to_sftp_operator.py
 ./airflow/contrib/operators/sagemaker_base_operator.py
