+++ tests/providers/amazon/aws/operators/test_s3_list.py
@@ -21,7 +21,7 @@ import unittest
 
 import mock
 
-from airflow.contrib.operators.s3_list_operator import S3ListOperator
+from airflow.providers.amazon.aws.operators.s3_list import S3ListOperator
 
 TASK_ID = 'test-s3-list-operator'
 BUCKET = 'test-bucket'
@@ -31,7 +31,7 @@ MOCK_FILES = ["TEST1.csv", "TEST2.csv", "TEST3.csv"]
 
 
 class TestS3ListOperator(unittest.TestCase):
-    @mock.patch('airflow.contrib.operators.s3_list_operator.S3Hook')
+    @mock.patch('airflow.providers.amazon.aws.operators.s3_list.S3Hook')
     def test_execute(self, mock_hook):
 
         mock_hook.return_value.list_keys.return_value = MOCK_FILES
